+++ packages/apps-config/src/endpoints/testingRelayPaseo.ts
@@ -3,8 +3,10 @@
 
 import type { EndpointOption } from './types.js';
 
+import { chainsFrequencyPaseoSVG, chainsPaseoBgPNG, chainsPaseoPNG } from '@polkadot/apps-config/ui/logos/chains';
+import { nodesIntegriteeSVG } from '@polkadot/apps-config/ui/logos/nodes';
+
 import { PASEO_GENESIS } from '../api/constants.js';
-import { chainsPaseoPNG } from '../ui/logos/chains/index.js';
 // import { testnetParachainSVG } from '../ui/logos/nodes/index.js';
 import { getTeleports } from './util.js';
 
@@ -29,6 +31,32 @@ export const testParasPaseo: Omit<EndpointOption, 'teleport'>[] = [
   //     logo: testnetParachainSVG
   //   }
   // }
+  {
+    homepage: 'https://www.frequency.xyz',
+    info: 'Frequency',
+    paraId: 4000,
+    providers: {
+      Integritee: 'wss://0.rpc.testnet.amplica.io'
+    },
+    text: 'Frequency',
+    ui: {
+      color: '#19455E',
+      logo: chainsFrequencyPaseoSVG
+    }
+  },
+  {
+    homepage: 'https://integritee.network',
+    info: 'integritee',
+    paraId: 2039,
+    providers: {
+      Integritee: 'wss://paseo.api.integritee.network'
+    },
+    text: 'Integritee Network (Paseo)',
+    ui: {
+      color: '#658ea9',
+      logo: nodesIntegriteeSVG
+    }
+  }
 ];
 
 export const testParasPaseoCommon: EndpointOption[] = [
@@ -58,6 +86,7 @@ export const testRelayPaseo: EndpointOption = {
   ],
   providers: {
     Amforc: 'wss://paseo.rpc.amforc.com',
+    'IBP-GeoDNS1': 'wss://rpc.ibp.network/paseo',
     'IBP-GeoDNS2': 'wss://rpc.dotters.network/paseo'
     // 'light client': 'light://substrate-connect/paseo'
   },