+++ packages/apps/src/overlays/LocalFork.tsx
@@ -2,7 +2,8 @@
 // SPDX-License-Identifier: Apache-2.0
 
 import React from 'react';
-import store from 'store';
+
+import { useApi } from '@polkadot/react-hooks';
 
 import { useTranslation } from '../translate.js';
 import BaseOverlay from './Base.js';
@@ -13,8 +14,9 @@ interface Props {
 
 function LocalFork ({ className }: Props): React.ReactElement<Props> | null {
   const { t } = useTranslation();
+  const { isLocalFork } = useApi();
 
-  if (store.get('isLocalFork')) {
+  if (isLocalFork) {
     return (
       <BaseOverlay
         className={className}