+++ packages/page-nfts/src/types.ts
@@ -5,7 +5,7 @@ import type { AccountId } from '@polkadot/types/interfaces';
 import type { PalletUniquesCollectionDetails, PalletUniquesCollectionMetadata } from '@polkadot/types/lookup';
 import type { BN } from '@polkadot/util';
 
-export interface CollectionSupportedIpfsData {
+export interface CollectionSupportedMetadata {
   name: string | null;
   image: string | null;
 }
@@ -19,7 +19,7 @@ export interface CollectionInfo {
   isOwnerMe: boolean;
   key: string;
   metadata: PalletUniquesCollectionMetadata | null;
-  ipfsData: CollectionSupportedIpfsData | null;
+  ipfsData: CollectionSupportedMetadata | null;
 }
 
 export interface CollectionInfoComplete extends CollectionInfo {