+++ CHANGELOG.md
@@ -1,5 +1,158 @@
 # CHANGELOG
 
+## 0.135.1 Mar 21, 2024
+
+Contributed:
+
+- Add integritee para to paseo (https://github.com/brenzi)
+- Polimec: add new RPC provider and change color (https://github.com/lrazovic)
+- Remove the Turing Network endpoint provided by the provider Dwellir (https://github.com/imstar15)
+- Update ethereumChains.ts (https://github.com/dastanfv)
+- Remove Automata 1RPC (https://github.com/laboon)
+- Add XCM encoder utility (https://github.com/serban300)
+- Restore Virto provider for Kreivo in Kusama (https://github.com/olanod)
+- Disable polkaholic (https://github.com/sourabhniyogi)
+
+Changes:
+
+- Bump @polkadot/* lib deps
+- Disable unreachable endpoints
+- Fix Collator call for compatibility
+
+
+## 0.134.1 Mar 14, 2024
+
+Contributed:
+
+- Update chainflip-perseverance archive endpoint
+- Change Xode RPC address (https://github.com/hgminerva)
+- Update translation.json (https://github.com/Marvel63)
+- Peerplays types bundle (https://github.com/serkixenos)
+- Add Paseo's IBP endpoint (https://github.com/miloskriz)
+- Add harmonie network (allfeat testnet) (https://github.com/SailorSnoW)
+- Wake up Integritee shell on Polkadot and adapt naming (https://github.com/brenzi)
+- Fix typo in condition check for a curve type (https://github.com/rzadp)
+- Changed metaquity to phyken network (https://github.com/Ekaanth)
+- Add dwellir liberland endpoint (https://github.com/gflow33)
+- Add Invo Network to Rococo (https://github.com/ayushmishra2005)
+- Add metadata for integritee crowdloan (https://github.com/brenzi)
+- Re-enable Edgeware Beresheet (https://github.com/JelliedOwl)
+- Add Xode Project (https://github.com/hgminerva)
+- Add Hyperbridge networks (https://github.com/seunlanlege)
+- Upgrade chopsticks version (https://github.com/qiweiii)
+- Adding RockX endpoint (https://github.com/AgoL5866)
+- Add Continuum Metaverse.Network to Polkadot Relay (https://github.com/tritm-cs)
+- Fix make local fork non persistent across sessions (https://github.com/qiweiii)
+- Add dwellir endpoint for robonomics (https://github.com/gflow33)
+- Fix switch rpc in url should not auto use local fork (https://github.com/qiweiii)
+- Fix updating the democracy proposal modal (https://github.com/Dudleyneedham)
+- Enable LuckyFriday Bridgehub (https://github.com/paradox-tt)
+- Support ipfs://ipfs/ links (https://github.com/jsidorenko)
+- Add logo and background for Paseo testnet (https://github.com/al3mart)
+- Adding RadiumBlock as Polkadot AssetHub, Kusama AssetHub, BridgeHub-Kusama and BridgeHub-Polkadot Endpoint provider (https://github.com/radiumb)
+- Add Kippu rpc endpoint for Kreivo (https://github.com/pandres95)
+- Integrate chopsticks (https://github.com/qiweiii)
+- Update 3dpass provider endpoint (https://github.com/fedosov)
+- Re-enable JelliedOwl Edgeware (https://github.com/JelliedOwl)
+- Disable encointer & westend endpoints (no funding) (https://github.com/stakeworld)
+- Add metadata for NeuroWeb crowdloan (https://github.com/NZT48)
+- Add Metadata for Polkadex crowdloan (https://github.com/ap211unitech)
+- Add Curio Kusama chain (https://github.com/alexxxkny)
+- Add Liberland network (https://github.com/DorianSternVukotic)
+- Reorganise Edgeware RPCs (https://github.com/JelliedOwl)
+- Enable Aventus provider (https://github.com/nahuseyoum)
+- Add coretime-westend RPC (https://github.com/BulatSaif)
+- Add Logion parachain endpoints (https://github.com/gdethier)
+- Re-enable hydradx rococo endpoint (https://github.com/jak-pan)
+- Adding RadiumBlock as Moonbeam Endpoint Provider (https://github.com/radiumb)
+- Fix dropdown options not visible for metadata upgradable extensions 
+- Adding RadiumBlock as Phala and Moonriver Endpoint provider (https://github.com/radiumb)
+- Add IBP endpoints to Paseo (https://github.com/senseless)
+- Add muse network (https://github.com/Moliholy)
+- Add dwellir endpoint for subsocial (https://github.com/gflow33)
+- Reactivate integritee rpc endpoints (https://github.com/brenzi)
+- Use archive nodes for SORA Kusama and Polkadot parachains (https://github.com/stefashkaa)
+- Integrate Laos Chain as Ethereum-Based Chain (https://github.com/asiniscalchi)
+- Add Paseo testnet (https://github.com/al3mart)
+- Remove Crab Dwellir endpoint (https://github.com/JayJay1024)
+- Update Logion RPCs (https://github.com/gdethier)
+- OpenZeppelin runtime template (https://github.com/KitHat)
+- Add people parachain RPC (https://github.com/BulatSaif)
+- Dwellir: update Polkadex parachain, add Polkadex mainnet (https://github.com/jakobilobi)
+- Update Dwellir productionRelayPolkadot.ts (https://github.com/Maharacha)
+- Dwellir Bifrost Kusama endpoint working again (https://github.com/jonathanudd)
+- Add SORA Polkadot Parachain node (https://github.com/stefashkaa)
+- Add edgscan.ink explorer (https://github.com/ShankarWarang)
+- Add Yerba Network to Kusama (https://github.com/agryaznov)
+- Adding RadiumBlock as Khala Endpoint provider (https://github.com/radiumb)
+- Return DAO IPCI endpoint (https://github.com/falconexe)
+- Rename OriginTrail Prachain to NeuroWeb (https://github.com/NZT48)
+- Update Dwellir productionRelayPolkadot.ts (https://github.com/benjichat)
+- Remove Logion para testnet (https://github.com/gdethier)
+- Update Curio Parachain (https://github.com/alexxxkny)
+- Updated md5 rococo parachain collator 1 provider url (https://github.com/sebastianmontero)
+- Ternoa is a live network (https://github.com/soufiane-capsule)
+- Add joyutils RPC endpoint for Joystream (https://github.com/kdembler)
+- Restore blockops rpcs endpoint (https://github.com/Damola12345)
+- Change logo and color of Energy Web X network (https://github.com/renesy)
+- Add RadiumBlock endpoint to Watr - Update productionRelayPolkadot.ts (https://github.com/radiumb)
+- Fixed urls for genshiro & equilibrium (https://github.com/mikolajsobolewski)
+- Add Helikon public RPC node for Polimec (https://github.com/kukabi)
+- Added Helikon RPC node endpoints for Phala, Khala and HydraDX (https://github.com/kukabi)
+- Integration of K-LAOS kusama parachain (https://github.com/asiniscalchi)
+- Set up Vara inflation pararms (https://github.com/EugenWay)
+- Add Chainflip networks (https://github.com/GabrielBuragev)
+- OnFinality Frequency Recovery (https://github.com/CrazySurfBoy)
+- Integrate LAOS Rococo Parachain (Giedi) (https://github.com/asiniscalchi)
+- Reenable Mangata's project rococo endpoint (https://github.com/tenequm)
+- Update tangle testnet rpc endpoint (https://github.com/salman01zp)
+- Adding RadiumBlock as Shiden Endpoint provider (https://github.com/radiumb)
+- Update productionRelayPolkadot.ts (https://github.com/benjichat)
+- Update parallel rpc (https://github.com/GopherJ)
+- Add Polimec Network (https://github.com/vstam1)
+- Reenable nodes (https://github.com/alex-nax)
+- Fix changelog typos (https://github.com/dzizazda)
+- Add public RPC endpoint for Energy Web X (https://github.com/renesy)
+- Add VTB Network Mainnet (https://github.com/Wasif-Jamal)
+- Octopus Provider for Myriad and Debio (https://github.com/RiXelanya)
+- Get isEthereum from chain properties (https://github.com/fgamundi)
+- Fix Expected object keys to be in ascending order (https://github.com/CrazySurfBoy)
+- Add Ternoa websocket (https://github.com/soufiane-capsule)
+- Fix make transfer example (https://github.com/ugur-eren)
+- Add Vara Testnet (https://github.com/EugenWay)
+- Fix transfer button disappearance (https://github.com/pacoyang)
+- Change ethereum icons based on URL (https://github.com/wirednkod)
+- Update overrides (https://github.com/Tbaut)
+- Fix typos (https://github.com/shuoer86)
+- Update kodadot link & logo (https://github.com/roiLeo)
+- Update @frequency/api-augment to v1.9.0 (https://github.com/shannonwells)
+- Update vara logo (https://github.com/EugenWay)
+- Update Litentry rococo-parachain endpoint (https://github.com/Kailai-Wang)
+- Remove Phala PoC5 since it going to take down (https://github.com/Leechael)
+- Removing Kapex RPC for Dwellir | Update productionRelayPolkadot.ts (https://github.com/Omgzilla)
+- Disable RadiumBlock assethub endpoints for kusama and polkadot (https://github.com/radiumb)
+- Add masverse node (https://github.com/duonghb53)
+- Add rococo coretime parachain rpc (https://github.com/BulatSaif)
+- Added LuckyFriday's Acala node (https://github.com/paradox-tt)
+- Added Dwellir endpoint for Bifrost Polkadot (https://github.com/gflow33)
+- Add Innovatorchain (https://github.com/masterdubs)
+- Add Subspace Gemini-3g Nova testing rpc (https://github.com/vedhavyas)
+- Update krest network providers (https://github.com/lavish0000)
+- Add innovatorchain in evm chains (https://github.com/masterdubs)
+- Remove Parity's Polkadot and Kusama RPC providers (https://github.com/PierreBesson)
+
+Changes
+
+- Disable unreachable endpoints (https://github.com/jacogr)
+- Bump yarn berry
+- Bump @polkadot/* deps
+- Disable staking for unknown rewards handling
+- Bump 3rd-party deps
+- Disable claims (non-working)
+- Support requestStatusFor for usePreimage
+- Fix break in backwards compatibility with preimages
+
+
 ## 0.133.1 Nov 19, 2023
 
 Contributed: